+++ NEWS.md
@@ -10,6 +10,8 @@
 - XFT traits for vulnerability curves
 - Recording of leaf PLC for all sub-models
 - Richard-type soil water movement
+- Soil hydrology revised
+- New functions for spwb in agricultural lands
 
 # Version 3.2.0
 - Correction of a bug arisen in old Rcpp versions